# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Crewline",
    "description": "Backend for the autonomous engineering agent platform",
    "version": "0.1.0"
  },
  "paths": {
...
    "/tickets": {
      "post": {
        "tags": [
          "tickets"
        ],
        "summary": "Create Ticket Route",
        "operationId": "create_ticket_route_tickets_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TicketCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TicketRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "tickets"
        ],
        "summary": "List Tickets Route",
        "operationId": "list_tickets_route_tickets_get",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/TicketStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "owner_agent_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Owner Agent Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Paginated_TicketRead_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/tickets/{ticket_id}": {
      "get": {
        "tags": [
          "tickets"
        ],
        "summary": "Get Ticket Route",
        "operationId": "get_ticket_route_tickets__ticket_id__get",
        "parameters": [
          {
            "name": "ticket_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Ticket Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TicketRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "tickets"
        ],
        "summary": "Update Ticket Route",
        "operationId": "update_ticket_route_tickets__ticket_id__patch",
        "parameters": [
          {
            "name": "ticket_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Ticket Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TicketUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TicketRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/mailbox": {
      "get": {
        "tags": [
          "mailbox"
        ],
        "summary": "Get Mailbox",
        "operationId": "get_mailbox_mailbox_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MailboxMessageRead"
                  },
                  "title": "Response Get Mailbox Mailbox Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/mailbox/messages": {
      "post": {
        "tags": [
          "mailbox"
        ],
        "summary": "Send Message",
        "operationId": "send_message_mailbox_messages_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "oneOf": [
                  {
                    "$ref": "#/components/schemas/SendRpcRequest"
                  },
                  {
                    "$ref": "#/components/schemas/SendRpcResponse"
                  },
                  {
                    "$ref": "#/components/schemas/SendNotification"
                  }
                ],
                "title": "Payload",
                "discriminator": {
                  "propertyName": "type",
                  "mapping": {
                    "rpc_request": "#/components/schemas/SendRpcRequest",
                    "rpc_response": "#/components/schemas/SendRpcResponse",
                    "notification": "#/components/schemas/SendNotification"
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MailboxMessageRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/mailbox/messages/{message_id}/ack": {
      "post": {
        "tags": [
          "mailbox"
        ],
        "summary": "Ack Message",
        "operationId": "ack_message_mailbox_messages__message_id__ack_post",
        "parameters": [
          {
            "name": "message_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Message Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/mailbox/messages/{message_id}/reject": {
      "post": {
        "tags": [
          "mailbox"
        ],
        "summary": "Reject Message Route",
        "operationId": "reject_message_route_mailbox_messages__message_id__reject_post",
        "parameters": [
          {
            "name": "message_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Message Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RejectMessageRequest"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MailboxMessageRead": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "type": {
            "$ref": "#/components/schemas/MessageType"
          },
          "sender": {
            "type": "string",
            "title": "Sender"
          },
          "recipient": {
            "type": "string",
            "title": "Recipient"
          },
          "payload": {
            "additionalProperties": true,
            "type": "object",
            "title": "Payload"
          },
          "requires_response": {
            "type": "boolean",
            "title": "Requires Response"
          },
          "correlation_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Correlation Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "type",
          "sender",
          "recipient",
          "payload",
          "requires_response",
          "correlation_id",
          "created_at"
        ],
        "title": "MailboxMessageRead"
      },
...
      "MessageType": {
        "type": "string",
        "enum": [
          "ticket_assigned",
          "ticket_unassigned",
          "ticket_cancelled",
          "ticket_review_requested",
          "rpc_request",
          "rpc_response",
          "notification"
        ],
        "title": "MessageType"
      },
      "NotificationPayload": {
        "properties": {
          "subject": {
            "type": "string",
            "title": "Subject"
          },
          "body": {
            "type": "string",
            "title": "Body"
          },
          "ticket_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ticket Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "subject",
          "body"
        ],
        "title": "NotificationPayload"
      },
      "Paginated_TicketRead_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/TicketRead"
            },
            "type": "array",
            "title": "Items"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          },
          "has_more": {
            "type": "boolean",
            "title": "Has More"
          }
        },
        "type": "object",
        "required": [
          "items",
          "limit",
          "offset",
          "has_more"
        ],
        "title": "Paginated[TicketRead]"
      },
      "RejectMessageRequest": {
        "properties": {
          "reason": {
            "type": "string",
            "minLength": 1,
            "title": "Reason"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "reason"
        ],
        "title": "RejectMessageRequest"
      },
      "RpcOutcome": {
        "type": "string",
        "enum": [
          "answered",
          "declined"
        ],
        "title": "RpcOutcome"
      },
      "RpcRequestPayload": {
        "properties": {
          "subject": {
            "type": "string",
            "title": "Subject"
          },
          "body": {
            "type": "string",
            "title": "Body"
          },
          "ticket_id": {
            "type": "integer",
            "title": "Ticket Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "subject",
          "body",
          "ticket_id"
        ],
        "title": "RpcRequestPayload"
      },
      "RpcResponsePayload": {
        "properties": {
          "body": {
            "type": "string",
            "title": "Body"
          },
          "outcome": {
            "$ref": "#/components/schemas/RpcOutcome"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "body",
          "outcome"
        ],
        "title": "RpcResponsePayload"
      },
      "SendNotification": {
        "properties": {
          "type": {
            "type": "string",
            "const": "notification",
            "title": "Type"
          },
          "recipient": {
            "type": "string",
            "title": "Recipient"
          },
          "payload": {
            "$ref": "#/components/schemas/NotificationPayload"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "type",
          "recipient",
          "payload"
        ],
        "title": "SendNotification"
      },
      "SendRpcRequest": {
        "properties": {
          "type": {
            "type": "string",
            "const": "rpc_request",
            "title": "Type"
          },
          "recipient": {
            "type": "string",
            "title": "Recipient"
          },
          "payload": {
            "$ref": "#/components/schemas/RpcRequestPayload"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "type",
          "recipient",
          "payload"
        ],
        "title": "SendRpcRequest"
      },
      "SendRpcResponse": {
        "properties": {
          "type": {
            "type": "string",
            "const": "rpc_response",
            "title": "Type"
          },
          "recipient": {
            "type": "string",
            "title": "Recipient"
          },
          "correlation_id": {
            "type": "integer",
            "title": "Correlation Id"
          },
          "payload": {
            "$ref": "#/components/schemas/RpcResponsePayload"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "type",
          "recipient",
          "correlation_id",
          "payload"
        ],
        "title": "SendRpcResponse"
      },
...
      "TicketCreate": {
        "properties": {
          "title": {
            "type": "string",
            "title": "Title"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "owner_agent_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Owner Agent Id"
          },
          "metadata": {
            "additionalProperties": true,
            "type": "object",
            "title": "Metadata"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "title"
        ],
        "title": "TicketCreate"
      },
...
      "TicketRead": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "status": {
            "$ref": "#/components/schemas/TicketStatus"
          },
          "owner_agent_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Owner Agent Id"
          },
          "created_by": {
            "type": "string",
            "title": "Created By"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "metadata": {
            "additionalProperties": true,
            "type": "object",
            "title": "Metadata"
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "description",
          "status",
          "owner_agent_id",
          "created_by",
          "created_at",
          "updated_at",
          "metadata"
        ],
        "title": "TicketRead"
      },
      "TicketStatus": {
        "type": "string",
        "enum": [
          "TODO",
          "IN_PROGRESS",
          "BLOCKED",
          "READY_FOR_QA",
          "IN_QA",
          "QA_FAILED",
          "DONE",
          "CANCELLED"
        ],
        "title": "TicketStatus"
      },
...
      "TicketUpdate": {
        "properties": {
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "metadata": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Metadata"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "TicketUpdate"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}